menu_app/requirements.txt
asgiref==3.6.0
Django==4.2
django-environ==0.10.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-mptt-admin==2.5.0
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
